+++ HISTORY.rst
@@ -2,7 +2,11 @@
 History
 =======
 
+0.2.0 (2021-05-24)
+------------------
+* Added 'setup' functionality to mozvpn
+* Extended documentation, added installation docs
+
 0.1.0 (2021-05-06)
 ------------------
-
 * First commit on github.